requirements.txt
flask>=2.3
requests>=2.28
Pillow>=9.0
